% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knobs.R
\name{plot}
\alias{plot}
\alias{plot,knobs}
\alias{plot,knobs,missing-method}
\alias{plot,knobs,formula-method}
\title{Plot method for mrgsims objects}
\usage{
plot(x, y, ...)

\S4method{plot}{knobs,missing}(x, yval = variables(x), auto.key = list(), mincol = 3, ...)

\S4method{plot}{knobs,formula}(
  x,
  y,
  show.grid = TRUE,
  lwd = 2,
  type = "l",
  yval = variables(x),
  auto.key = list(columns = 1),
  scales = list(y = list(relation = "free")),
  ...
)
}
\arguments{
\item{x}{mrgsims object}

\item{y}{a formula passed to xyplot}

\item{...}{arguments passed to xyplot}

\item{yval}{y variables to plot}

\item{auto.key}{passed to xyplot}

\item{mincol}{minimum number of columns in key}

\item{show.grid}{print grid in the plot}

\item{lwd}{passed to xyplot}

\item{type}{passed to xyplot}

\item{scales}{passed to xyplot}
}
\description{
Plot method for mrgsims objects
}
